Improvements on the container build
This change pulls up the base image update, so that this can be better
cached. It also allows to set and empty KURA_COMMIT, so that the most
recent commit on this branch may be used. It adds a chmod a+x to the
content of /usr/local/bin, just in case.